refactored code with lib folder
