+++ server.js
@@ -1,19 +1,17 @@
-const logger = require('./index');
+const dojotLibs = require('./index');
 const express = require('express');
 
-var app = express();
+//defining log
+var logger = dojotLibs.logger;
 
-app.get('/setLog', (req, res) => {
-    res.send("Level of debugger is set to " + req.query.level);
-    logger.transports[0].level = req.query.level;
-    logger.debug("debug");
-    logger.info("info");
-    logger.warn("warn");
-    logger.error("error");
-}, e => {
-      res.status(400).send(e);
-  });
+//defining app express
+const app = express();
 
-  app.listen(3000, () => {
-    console.log(`Started up at port 3000`);
-  });
+//setting log debug route to app
+dojotLibs.loggerRoute(app);
+
+
+
+app.listen(3000, () => {
+console.log(`Started up at port 3000`);
+});
